# Conflicts:
#	marketing-cs/src/main/java/com/br/marketing/client/rulecleaning/CleanConfigDTO.java
#	marketing-cs/src/main/java/com/br/marketing/service/Impl/PushRuleServiceImpl.java
#	marketing-cs/src/main/java/com/br/marketing/service/PushRuleService.java
#	marketing-cs/src/main/java/com/br/marketing/service/clean/common/DataCleanService.java
#	marketing-cs/src/main/java/com/br/marketing/service/clean/common/impl/DataCleanServiceImpl.java
#	marketing-cs/src/main/java/com/br/marketing/service/ruleCleaning/impl/RuleCleaningServiceImpl.java
#	marketing-inner-api/src/main/java/com/br/marketing/innerapi/controller/RuleCleaningController.java

+++ marketing-cs/src/main/java/com/br/marketing/mapper/MarketingDataCleanGeneralFieldConfigMapperBase.java
@@ -2,11 +2,12 @@ package com.br.marketing.mapper;
 
 import com.br.marketing.entity.MarketingDataCleanGeneralFieldConfig;
 import com.br.marketing.entity.MarketingDataCleanGeneralFieldConfigExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface MarketingDataCleanGeneralFieldConfigMapperBase {
-    long countByExample(MarketingDataCleanGeneralFieldConfigExample example);
+    int countByExample(MarketingDataCleanGeneralFieldConfigExample example);
 
     int deleteByExample(MarketingDataCleanGeneralFieldConfigExample example);
